Rename Aliases.hpp to types.hpp.
