+++ src/log_surgeon/finite_automata/NfaState.hpp
@@ -13,10 +13,10 @@
 
 #include <fmt/format.h>
 
-#include <log_surgeon/Aliases.hpp>
 #include <log_surgeon/finite_automata/StateType.hpp>
 #include <log_surgeon/finite_automata/TaggedTransition.hpp>
 #include <log_surgeon/finite_automata/UnicodeIntervalTree.hpp>
+#include <log_surgeon/types.hpp>
 
 namespace log_surgeon::finite_automata {
 template <StateType state_type>
